reduce size and fix some bugs
